Rename Interleaver --> StreamCollator
